pyproject.toml
[tool.poetry]
name = "lunabot"
version = "1.0.0"
description = "Bot for Midnight in Manhattan"
authors = ["Skye Im <skye.im@nyu.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.4"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
